Fix Paths/Skipmaster/Ability_WorldTeleport.cs
The function DrawHighlight() doesn't work for some reason, but using GizmoUpdateOnMouseover() works just fine.
I don't understand what CanHitTargetTile() is suppose to do, when "Ability_WorldTeleportNight" straight up overrides the value depending on the night cycle.
My code for CanHitTargetTile() checks is the mouse in range of the ability and prevents teleporting to the current location (it can happen with night skip now because of the override).
Also for a balancing issue. I think that range 9999 for night skip is way too overpowered, since the player can jump right into the ship with it.
Also, odd thing, the far skip ability allows teleporting only to the home location (colony). My code fixes that.
Cheers guys :)